renaming server main file
